rpc: move LoadSecurityOptions and PGURL to a new lightweight package
This allows clarifies their scope as separate from the RPC logic.

Release note: None
