+++ pkg/cli/BUILD.bazel
@@ -133,6 +133,7 @@ go_library(
         "//pkg/rpc",
         "//pkg/security",
         "//pkg/security/certnames",
+        "//pkg/security/clientsecopts",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
         "//pkg/security/username",
