+++ pkg/cli/start.go
@@ -37,6 +37,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/rpc"
+	"github.com/cockroachdb/cockroach/pkg/security/clientsecopts"
 	"github.com/cockroachdb/cockroach/pkg/security/username"
 	"github.com/cockroachdb/cockroach/pkg/server"
 	"github.com/cockroachdb/cockroach/pkg/server/status"
@@ -519,7 +520,7 @@ func runStart(cmd *cobra.Command, args []string, startSingleNode bool) (returnEr
 			// earlier (e.g. above, in the runStart function) because
 			// at this time the address and port have not been resolved yet.
 			clientConnOptions, serverParams := makeServerOptionsForURL(&serverCfg)
-			pgURL, err := rpc.PGURL(clientConnOptions, serverParams, url.User(username.RootUser))
+			pgURL, err := clientsecopts.PGURL(clientConnOptions, serverParams, url.User(username.RootUser))
 			if err != nil {
 				log.Errorf(ctx, "failed computing the URL: %v", err)
 				return
@@ -916,12 +917,12 @@ func waitForShutdown(
 // is finalized late in the network initialization sequence.
 func makeServerOptionsForURL(
 	serverCfg *server.Config,
-) (rpc.ClientConnectOptions, rpc.ServerParameters) {
-	clientConnOptions := rpc.ClientConnectOptions{
+) (clientsecopts.ClientOptions, clientsecopts.ServerParameters) {
+	clientConnOptions := clientsecopts.ClientOptions{
 		Insecure: serverCfg.Config.Insecure,
 		CertsDir: serverCfg.Config.SSLCertsDir,
 	}
-	serverParams := rpc.ServerParameters{
+	serverParams := clientsecopts.ServerParameters{
 		ServerAddr:      serverCfg.Config.SQLAdvertiseAddr,
 		DefaultPort:     base.DefaultPort,
 		DefaultDatabase: catalogkeys.DefaultDatabaseName,
@@ -955,7 +956,7 @@ func reportServerInfo(
 	// earlier (e.g. above, in the runStart function) because
 	// at this time the address and port have not been resolved yet.
 	clientConnOptions, serverParams := makeServerOptionsForURL(serverCfg)
-	pgURL, err := rpc.PGURL(clientConnOptions, serverParams, url.User(username.RootUser))
+	pgURL, err := clientsecopts.PGURL(clientConnOptions, serverParams, url.User(username.RootUser))
 	if err != nil {
 		log.Ops.Errorf(ctx, "failed computing the URL: %v", err)
 		return err
