+++ pkg/rpc/BUILD.bazel
@@ -17,7 +17,6 @@ go_library(
         "heartbeat.go",
         "keepalive.go",
         "metrics.go",
-        "pg.go",
         "snappy.go",
         "tls.go",
     ],
@@ -31,9 +30,7 @@ go_library(
         "//pkg/roachpb",
         "//pkg/security",
         "//pkg/security/certnames",
-        "//pkg/security/securityassets",
         "//pkg/security/username",
-        "//pkg/server/pgurl",
         "//pkg/settings/cluster",
         "//pkg/util/contextutil",
         "//pkg/util/envutil",
@@ -44,7 +41,6 @@ go_library(
         "//pkg/util/log/severity",
         "//pkg/util/metric",
         "//pkg/util/netutil",
-        "//pkg/util/netutil/addr",
         "//pkg/util/stop",
         "//pkg/util/syncutil",
         "//pkg/util/timeutil",
