+++ pkg/security/clientsecopts/BUILD.bazel
@@ -1,0 +1,18 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "clientsecopts",
+    srcs = [
+        "client_opts.go",
+        "doc.go",
+    ],
+    importpath = "github.com/cockroachdb/cockroach/pkg/security/clientsecopts",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/security/certnames",
+        "//pkg/security/securityassets",
+        "//pkg/security/username",
+        "//pkg/server/pgurl",
+        "//pkg/util/netutil/addr",
+    ],
+)
