+++ pkg/server/BUILD.bazel
@@ -118,6 +118,7 @@ go_library(
         "//pkg/scheduledjobs",
         "//pkg/security",
         "//pkg/security/certnames",
+        "//pkg/security/clientsecopts",
         "//pkg/security/password",
         "//pkg/security/securityassets",
         "//pkg/security/username",
