+++ pkg/server/server_sql.go
@@ -40,6 +40,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/rpc"
 	"github.com/cockroachdb/cockroach/pkg/rpc/nodedialer"
 	"github.com/cockroachdb/cockroach/pkg/scheduledjobs"
+	"github.com/cockroachdb/cockroach/pkg/security/clientsecopts"
 	"github.com/cockroachdb/cockroach/pkg/security/username"
 	"github.com/cockroachdb/cockroach/pkg/server/diagnostics"
 	"github.com/cockroachdb/cockroach/pkg/server/pgurl"
@@ -676,11 +677,11 @@ func newSQLServer(ctx context.Context, cfg sqlServerArgs) (*SQLServer, error) {
 		sqlExecutorTestingKnobs = sql.ExecutorTestingKnobs{}
 	}
 
-	ccopts := rpc.ClientConnectOptions{
+	ccopts := clientsecopts.ClientOptions{
 		Insecure: cfg.Config.Insecure,
 		CertsDir: cfg.Config.SSLCertsDir,
 	}
-	sparams := rpc.ServerParameters{
+	sparams := clientsecopts.ServerParameters{
 		ServerAddr:      cfg.Config.SQLAdvertiseAddr,
 		DefaultPort:     base.DefaultPort,
 		DefaultDatabase: catalogkeys.DefaultDatabaseName,
@@ -689,7 +690,7 @@ func newSQLServer(ctx context.Context, cfg sqlServerArgs) (*SQLServer, error) {
 	nodeInfo := sql.NodeInfo{
 		AdminURL: cfg.AdminURL,
 		PGURL: func(user *url.Userinfo) (*pgurl.URL, error) {
-			return rpc.PGURL(ccopts, sparams, user)
+			return clientsecopts.PGURL(ccopts, sparams, user)
 		},
 		LogicalClusterID: cfg.rpcContext.LogicalClusterID.Get,
 		NodeID:           cfg.nodeIDContainer,
